Refactor Flight Control to create two executables - One to run tests (GTest) - One to run the node
