+++ src/rov/flight_controller/src/main.cpp
@@ -1,0 +1,12 @@
+#include <rclcpp/rclcpp.hpp>
+#include <flight_controller/flight_controller.hpp>
+
+int main(int argc, char ** argv) {
+    rclcpp::init(argc, argv);
+    auto node = std::make_shared<FlightController>();
+    rclcpp::executors::MultiThreadedExecutor exec;
+    exec.add_node(node);
+    exec.spin();
+    rclcpp::shutdown();
+    return 0;
+}
