+++ src/rov/flight_controller/test/test_simple.cpp
@@ -1,0 +1,5 @@
+#include <gtest/gtest.h>
+
+TEST(flight_controller_tests, test_simple) {
+    EXPECT_TRUE(true);
+}
